level/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
click==8.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
discord==2.3.2
discord.py==2.5.2
distlib==0.3.9
dnspython==2.7.0
executing==2.2.0
filelock==3.18.0
Flask==3.0.0
flask-cors==5.0.1
fonttools==4.59.1
frozenlist==1.6.0
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pydantic==1.10.13
Pygments==2.19.1
pymongo==4.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==26.4.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.13.2
uvicorn==0.27.0
virtualenv==20.31.2
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.0
